Simplify tidy tests
