+++ xtask/tests/tidy.rs
@@ -1,11 +1,41 @@
-mod cli;
-
 use std::{
     collections::HashMap,
     path::{Path, PathBuf},
 };
 
-use xtask::{not_bash::fs2, project_root, rust_files};
+use xtask::{
+    codegen::{self, Mode},
+    not_bash::fs2,
+    project_root, run_rustfmt, rust_files,
+};
+
+#[test]
+fn generated_grammar_is_fresh() {
+    if let Err(error) = codegen::generate_syntax(Mode::Verify) {
+        panic!("{}. Please update it by running `cargo xtask codegen`", error);
+    }
+}
+
+#[test]
+fn generated_tests_are_fresh() {
+    if let Err(error) = codegen::generate_parser_tests(Mode::Verify) {
+        panic!("{}. Please update tests by running `cargo xtask codegen`", error);
+    }
+}
+
+#[test]
+fn generated_assists_are_fresh() {
+    if let Err(error) = codegen::generate_assists_docs(Mode::Verify) {
+        panic!("{}. Please update assists by running `cargo xtask codegen`", error);
+    }
+}
+
+#[test]
+fn check_code_formatting() {
+    if let Err(error) = run_rustfmt(Mode::Verify) {
+        panic!("{}. Please format the code by running `cargo format`", error);
+    }
+}
 
 #[test]
 fn rust_files_are_tidy() {
